Add separate file for config class.
