+++ lib/tp_cli.rb
@@ -1,6 +1,7 @@
 require 'typhoeus'
 require "tp_cli/version"
 require 'yaml'
+require 'configclass'
 
 module TpCommandLine
   class ActivityTrack
@@ -22,31 +23,9 @@ module TpCommandLine
       end
     end
 
-    def load_file(file)
-      if File.exists?(file)
-        return YAML.load(File.read(file))
-      else
-        file = {}
-      end
-    end
-
-    def load_config
-      hash1 = load_file(File.join(Dir.home, ".timepulse.yml"))
-      hash2 = load_file('config/.timepulse.yml')
-
-      config_hash = hash1.merge(hash2)
-
-      missing_fields = ['timepulse_url', 'project_id', 'login', 'authorization'].find_all {|k| !config_hash.keys.include? k}
-      unless missing_fields.empty?
-        exit "Missing necessary parameter/s: #{missing_fields.join(", ")}"
-      end
-
-      return config_hash
-    end
-
     #POST to create method in TP Activities controller; create an Activity in db
     def record_activity_in_TP(description)
-      config = load_config
+      config = TpCommandLine::Config.new.load_config
       request = Typhoeus::Request.new(
         config['timepulse_url'],
         method: :post,
